First typeclass implementation for Joda time. Package restructuring to reflect that.
