+++ src/main/scala/com/fortysevendeg/scalacheck/datetime/instances/j8.scala
@@ -1,0 +1,3 @@
+package com.fortysevendeg.scalacheck.datetime.instances
+
+trait J8Instances
