+++ src/main/scala/com/fortysevendeg/scalacheck/datetime/instances/package.scala
@@ -1,0 +1,6 @@
+package com.fortysevendeg.scalacheck.datetime
+
+package object instances {
+  object joda extends JodaInstances
+  object j8 extends J8Instances
+}
